+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 - Added ability to have a Vagrantfile.local file that will override settings in the checked in Vagrantfile
 - Added example drush alias file
 - Added post-provision example script that copies configuration files and drush aliases if they exist
-- Updated to 1.1.17 of puppet-forumone
+- Updated to 1.1.18 of puppet-forumone
 - Added configuration to example settings.vm.php to disable Google Analytics, Secure Pages and core search indexing
 - Updated max_execution_time for Drupal to 180 seconds
 - Updated example post-provision script to copy .htaccess if it exists