+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added configuration to example settings.vm.php to disable Google Analytics, Secure Pages and core search indexing
 - Updated max_execution_time for Drupal to 180 seconds
 - Updated example post-provision script to copy .htaccess if it exists
+- Added support for vagrant-cachier plugin
 
 ## 1.1.12 (9 May, 2014)
 - Changing meaning of roles. App is where Drupal is intended to live, so it symlinks settings, htaccess, etc. App is where Varnish is cleared. DB is where drush commands that affect the database are executed.