+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 - Added support for vagrant-cachier plugin
 - Fixed issue with copying drush aliases when they do not exist
 - Changed ".cap" to ".rake" to make more clear they are Rake tasks
+- Refactored tasks to allow for different platforms
+- Added documentation to Drupal platform add-on and tasks
+- Added build script tasks for rsync
+- Added tasks and platform for WordPress and WP-CLI
 
 ## 1.1.12 (9 May, 2014)
 - Changing meaning of roles. App is where Drupal is intended to live, so it symlinks settings, htaccess, etc. App is where Varnish is cleared. DB is where drush commands that affect the database are executed.